Feat/loc5 - Finished validate for CRUD MarcaController

+++ app/Http/Controllers/MarcaController.php
@@ -17,15 +17,17 @@ class MarcaController extends Controller
     public function index()
     {
         $marcas = $this->marca->all();
-        return $marcas;
+        return response()->json($marcas);
     }
     /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
+
+        $request->validate($this->marca->rules(), $this->marca->feedback());
         $marca = $this->marca->create($request->all());
-        return $marca;
+        return response()->json($marca, 201);
     }
 
     /**
@@ -34,7 +36,10 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
-        return $marca;
+        if($marca === null){
+            return response()->json(['erro' => 'Nenhum registro encontrado'], 404);
+        }
+        return response()->json($marca);
     }
 
     /**
@@ -42,9 +47,24 @@ class MarcaController extends Controller
      */
     public function update(Request $request, $id)
     {
+
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return response()->json(['erro' => 'Nenhum registro encontrado'], 404);
+        }
+        if($request->method() === 'PATCH'){
+            $dinamicRules = array();
+            foreach ($marca->rules() as $input => $rule){
+                if(array_key_exists($input, $request->all())){
+                    $dinamicRules[$input] = $rule;
+                }
+            }
+            $request->validate($dinamicRules, $marca->feedback());
+        }else{
+            $request->validate($marca->rules(), $marca->feedback());
+        }
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca);
     }
 
     /**
@@ -53,7 +73,10 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return response()->json(['erro' => 'Nenhum registro encontrado'], 404);
+        }
         $marca->delete();
-        return $marca;
+        return response()->json($marca);
     }
 }
